Criação de Pedido e PlacaDeVideo Atualizado

+++ src/main/java/br/unitins/tp1/placadevideo/dto/Request/ItemPedidoRequestDTO.java
@@ -1,8 +1,10 @@
 package br.unitins.tp1.placadevideo.dto.Request;
 
+import java.math.BigDecimal;
+
 public record ItemPedidoRequestDTO(
     Long idProduto, 
     Integer quantidade,
-    Double preco) {
+    BigDecimal preco) {
     
 }
