+++ src/main/java/br/unitins/tp1/placadevideo/dto/Response/ItemPedidoResponsetDTO.java
@@ -1,12 +1,14 @@
 package br.unitins.tp1.placadevideo.dto.Response;
 
-import br.unitins.tp1.placadevideo.model.ItemPedido;
+import java.math.BigDecimal;
+
+import br.unitins.tp1.placadevideo.model.pedido.ItemPedido;
 
 public record ItemPedidoResponsetDTO(
     Long idProduto,
     String nome, 
     Integer quantidade,
-    Double preco) {
+    BigDecimal preco) {
     
         public static ItemPedidoResponsetDTO valueOf(ItemPedido itemPedido){
             return new ItemPedidoResponsetDTO(
